chore: move profile-related pages out of the internal pages
