+++ front/src/pages/profile/experiences/ExperienceCreatePage.tsx
@@ -4,11 +4,11 @@ import * as React from 'react';
 import { ScrollView, StyleSheet, View } from 'react-native';
 import { Appbar } from 'react-native-paper';
 
-import ReferenceForm from '@/components/references/ReferenceForm';
-import { Reference } from '@/models/types';
-import { usePatchReferenceMutation } from '@/store/slice/api';
+import ExperienceForm from '@/components/experiences/ExperienceForm';
+import { Experience } from '@/models/types';
+import { usePostExperienceMutation } from '@/store/slice/api';
 
-import { ProfileStackParamList } from '../../ProfileNav';
+import { ProfileStackParamList } from '../ProfileNav';
 
 const styles = StyleSheet.create({
   container: {
@@ -51,32 +51,26 @@ const styles = StyleSheet.create({
   },
 });
 
-type ReferenceUpdatePageProps = NativeStackScreenProps<
+type ExperienceCreatePageProps = NativeStackScreenProps<
   ProfileStackParamList,
-  'ReferenceUpdate'
+  'ExperienceCreate'
 >;
 
-const ReferenceUpdatePage = ({
+const ExperienceCreatePage = ({
   route,
   navigation,
-}: ReferenceUpdatePageProps) => {
-  // Constants
-  const { id, firstName, lastName, address, email, phone, company } =
-    route.params as Reference;
-
+}: ExperienceCreatePageProps) => {
   // Hooks
 
   // Form State
   const [formData, setFormData] = useState({
-    id,
-    firstName: firstName ?? '',
-    lastName: lastName ?? '',
-    firstLine: address?.firstLine ?? '',
-    zipCode: address?.zipCode ?? '',
-    city: address?.city ?? '',
-    email: email ?? '',
-    phone: phone ?? '',
-    companyName: company?.name ?? '',
+    jobTitle: '',
+    firstLine: '',
+    zipCode: '',
+    city: '',
+    companyName: '',
+    startDate: null,
+    endDate: null,
   });
 
   // To set the action buttons in the appbar for saving the changes
@@ -91,33 +85,33 @@ const ReferenceUpdatePage = ({
   }, [navigation, formData]);
 
   // Api calls
-  const [patchReference] = usePatchReferenceMutation();
+  const [postExperience] = usePostExperienceMutation();
 
   // Methods
 
   const checkPressed = useCallback(() => {
-    const updatedReference: Partial<Reference> = {
-      id: formData.id,
+    const updatedExperience: Partial<Experience> = {
       company: {
         name: formData.companyName,
+      },
+      job: {
+        title: formData.jobTitle,
       },
       address: {
         firstLine: formData.firstLine,
         zipCode: formData.zipCode,
         city: formData.city,
       },
-      email: formData.email,
-      phone: formData.phone,
-      firstName: formData.firstName,
-      lastName: formData.lastName,
+      startDate: formData.startDate,
+      endDate: formData.endDate,
     };
 
-    patchReference(updatedReference)
+    postExperience(updatedExperience)
       .unwrap()
       .then((r) => {
         navigation.goBack();
       });
-  }, [formData, patchReference, navigation]);
+  }, [formData, postExperience, navigation]);
 
   return (
     <ScrollView
@@ -125,10 +119,10 @@ const ReferenceUpdatePage = ({
       contentContainerStyle={styles.contentContainer}
     >
       <View style={styles.verticalCenterContainer}>
-        <ReferenceForm formData={formData} setFormData={setFormData} />
+        <ExperienceForm formData={formData} setFormData={setFormData} />
       </View>
     </ScrollView>
   );
 };
 
-export default ReferenceUpdatePage;
+export default ExperienceCreatePage;
